64 bit compatibility

+++ src/ICalConverter.cpp
@@ -488,7 +488,7 @@ string ICalConverter::appendControlM(string szContents)
     std::istringstream iss(szContents);
     while (getline(iss, szLine)) {
 
-    unsigned int loc = szLine.length();
+    size_t loc = szLine.length();
 
     if (loc != 0) {
         szLine = szLine + "\r";
@@ -2861,7 +2861,7 @@ vector < string > ICalConverter::getRecurrenceDates(string strIcalComp,
             break;
     }
     if (strIcalComp.find(strRuleType) != string::npos ) {
-    unsigned int sPos = 0;
+    size_t sPos = 0;
     int count = 0;
     while ((sPos =
         strIcalComp.find(strRuleType, sPos)) != string::npos) {
